Разделяет main.js на модули. Добавляет модули data.js и util.js
